requirements.txt
gradio
pandas
numpy
tensorflow-cpu
scikit-learn
matplotlib
seaborn
plotly
openpyxl
